fix: comments made 1/06/2023 PR #38
