+++ crates/differ/src/apply_diff.rs
@@ -11,7 +11,7 @@ pub(crate) fn remove(start: usize, stop: usize, s: &str) -> String {
 	result
 }
 
-pub fn apply_diff(s: &str, diffs: &Box<[StringDiffOp]>) -> String {
+pub fn apply_diff(s: &str, diffs: Vec<StringDiffOp>) -> String {
 	let mut new_string: String = s.into();
 
 	for i in diffs.iter() {
@@ -67,20 +67,20 @@ mod tests {
 
 		assert_eq!(
 			String::from("sitting"),
-			super::apply_diff("kitten", &test_box)
+			super::apply_diff("kitten", test_box.to_vec())
 		);
 		assert_eq!(
 			String::from("Sunday"),
-			super::apply_diff("Saturday", &test_box_2)
+			super::apply_diff("Saturday", test_box_2.to_vec())
 		);
 		assert_eq!(
 			String::from("SETS"),
-			super::apply_diff("RESET", &test_box_3)
+			super::apply_diff("RESET", test_box_3.to_vec())
 		);
-		assert_eq!(String::from("heeoy"), super::apply_diff("hey", &test_box_4));
+		assert_eq!(String::from("heeoy"), super::apply_diff("hey", test_box_4.to_vec()));
 		assert_eq!(
 			String::from("skater"),
-			super::apply_diff("kate", &test_box_5)
+			super::apply_diff("kate", test_box_5.to_vec())
 		);
 	}
 }
